refactor(macros): split semantic pipeline modules
