+++ statum-macros/src/machine/emission/mod.rs
@@ -1,3 +1,4 @@
+mod builder;
 mod builders;
 mod presentation;
 mod support;
