+++ statum-macros/src/validators/emission/builders.rs
@@ -1,6 +1,11 @@
 use quote::{format_ident, quote};
 use syn::{Generics, Ident, ImplItem, Path, Type};
 
+use super::batch_finalization::{
+    generate_batch_finalization, BatchAsyncMode, BatchFieldSource,
+    BatchFinalizationOperation, BatchFinalizationPlan,
+};
+use super::into_machine::{generate_into_machine_builder, IntoMachineBuilderContext};
 use super::shared::{
     field_binding_tokens, field_storage_tokens, generic_usage_marker_tokens,
     machine_scoped_item_path, merged_where_clause_tokens,
@@ -40,44 +45,6 @@ pub(crate) struct ValidatorBuilderSurfaceContext<'a> {
     pub(crate) modified_methods: &'a [ImplItem],
     pub(crate) async_token: &'a proc_macro2::TokenStream,
     pub(crate) machine_vis: &'a syn::Visibility,
-}
-
-struct IntoMachineBuilderContext<'a> {
-    builder_ident: &'a Ident,
-    struct_ident: &'a Type,
-    machine_generics: &'a Generics,
-    machine_state_ty: &'a proc_macro2::TokenStream,
-    field_names: &'a [Ident],
-    field_types: &'a [Type],
-    validator_checks: &'a [proc_macro2::TokenStream],
-    validator_report_checks: &'a [proc_macro2::TokenStream],
-    async_token: &'a proc_macro2::TokenStream,
-    machine_vis: &'a syn::Visibility,
-}
-
-enum BatchFieldSource<'a> {
-    SharedAcrossItems {
-        field_builder_chain: &'a proc_macro2::TokenStream,
-    },
-    PerItemByFn {
-        field_builder_chain: &'a proc_macro2::TokenStream,
-    },
-}
-
-enum BatchAsyncMode {
-    Sync,
-    Async,
-}
-
-enum BatchFinalizationOperation {
-    Build,
-    BuildReport,
-}
-
-struct BatchFinalizationPlan<'a> {
-    operation: BatchFinalizationOperation,
-    async_mode: BatchAsyncMode,
-    field_source: BatchFieldSource<'a>,
 }
 
 pub(crate) fn validator_builder_surface(
@@ -188,95 +155,6 @@ pub(crate) fn validator_builder_surface(
 
         #into_machine_builder_impl
         #batch_builder_impl
-    }
-}
-
-fn generate_into_machine_builder(
-    context: IntoMachineBuilderContext<'_>,
-) -> proc_macro2::TokenStream {
-    let builder_ident = context.builder_ident;
-    let struct_ident = context.struct_ident;
-    let machine_generics = context.machine_generics;
-    let machine_state_ty = context.machine_state_ty;
-    let field_names = context.field_names;
-    let field_types = context.field_types;
-    let validator_checks = context.validator_checks;
-    let validator_report_checks = context.validator_report_checks;
-    let validator_report_count = validator_report_checks.len();
-    let async_token = context.async_token;
-    let machine_vis = context.machine_vis;
-    let extra_machine_generics = extra_generics(machine_generics);
-    let slot_storage_idents = slot_storage_idents(field_names.len());
-    let slot_state_idents = slot_state_idents(field_names.len());
-    let builder_defaults =
-        builder_generics(&extra_machine_generics, true, &slot_state_idents, true);
-    let complete_slots = slot_state_idents
-        .iter()
-        .map(|_| quote! { true })
-        .collect::<Vec<_>>();
-    let complete_builder_ty_generics = generic_argument_tokens(
-        extra_machine_generics.params.iter(),
-        Some(quote! { '__statum_row }),
-        &complete_slots,
-    );
-    let complete_builder_impl_generics_decl =
-        builder_generics(&extra_machine_generics, true, &[], false);
-    let (complete_builder_impl_generics, _, complete_builder_where_clause) =
-        complete_builder_impl_generics_decl.split_for_impl();
-
-    let struct_fields = field_storage_tokens(&slot_storage_idents, field_types);
-    let field_bindings = field_binding_tokens(field_names, &slot_storage_idents);
-    let setters = slot_setter_impls(
-        SlotSetterContext {
-            builder_ident,
-            machine_vis,
-            extra_machine_generics: &extra_machine_generics,
-            field_names,
-            field_types,
-            slot_state_idents: &slot_state_idents,
-            slot_storage_idents: &slot_storage_idents,
-            row_lifetime: Some(quote! { '__statum_row }),
-        },
-        |assignments| {
-            quote! {
-                #builder_ident {
-                    __statum_item: self.__statum_item,
-                    #(#assignments),*
-                }
-            }
-        },
-    );
-
-    quote! {
-        #[doc(hidden)]
-        #machine_vis struct #builder_ident #builder_defaults {
-            __statum_item: &'__statum_row #struct_ident,
-            #(#struct_fields),*
-        }
-
-        #(#setters)*
-
-        impl #complete_builder_impl_generics #builder_ident #complete_builder_ty_generics #complete_builder_where_clause {
-            #machine_vis #async_token fn build(self) -> core::result::Result<#machine_state_ty, statum::Error> {
-                let __statum_persisted = self.__statum_item;
-                #(#field_bindings)*
-                #(#validator_checks)*
-
-                Err(statum::Error::InvalidState)
-            }
-
-            #machine_vis #async_token fn build_report(self) -> statum::RebuildReport<#machine_state_ty> {
-                let __statum_persisted = self.__statum_item;
-                let mut __statum_attempts = ::std::vec::Vec::with_capacity(#validator_report_count);
-                #(#field_bindings)*
-                #(#validator_report_checks)*
-
-                statum::RebuildReport {
-                    attempts: __statum_attempts,
-                    result: Err(statum::Error::InvalidState),
-                }
-            }
-        }
     }
 }
 
@@ -521,72 +399,6 @@ pub(crate) fn batch_builder_implementation(
 
             #async_token fn __private_finalize_reports(self) -> Vec<statum::RebuildReport<#machine_state_ty>> {
                 #per_item_report_implementation
-            }
-        }
-    }
-}
-
-fn generate_batch_finalization(plan: BatchFinalizationPlan<'_>) -> proc_macro2::TokenStream {
-    let builder_method = match plan.operation {
-        BatchFinalizationOperation::Build => format_ident!("build"),
-        BatchFinalizationOperation::BuildReport => format_ident!("build_report"),
-    };
-
-    match (plan.async_mode, plan.field_source) {
-        (
-            BatchAsyncMode::Sync,
-            BatchFieldSource::SharedAcrossItems { field_builder_chain },
-        ) => {
-            quote! {
-                __statum_items
-                    .into_iter()
-                    .map(|__statum_item| {
-                        __statum_item.into_machine()
-                            #field_builder_chain
-                            .#builder_method()
-                    })
-                    .collect()
-            }
-        }
-        (
-            BatchAsyncMode::Async,
-            BatchFieldSource::SharedAcrossItems { field_builder_chain },
-        ) => {
-            quote! {
-                statum::__private::futures::future::join_all(
-                    __statum_items.iter().map(|__statum_item| {
-                        __statum_item.into_machine()
-                            #field_builder_chain
-                            .#builder_method()
-                    })
-                ).await
-            }
-        }
-        (BatchAsyncMode::Sync, BatchFieldSource::PerItemByFn { field_builder_chain }) => {
-            quote! {
-                let __statum_field_fn = self.__statum_fields_fn;
-                self.__statum_items
-                    .into_iter()
-                    .map(|__statum_item| {
-                        let __statum_fields = __statum_field_fn(&__statum_item);
-                        __statum_item.into_machine()
-                            #field_builder_chain
-                            .#builder_method()
-                    })
-                    .collect()
-            }
-        }
-        (BatchAsyncMode::Async, BatchFieldSource::PerItemByFn { field_builder_chain }) => {
-            quote! {
-                let __statum_field_fn = &self.__statum_fields_fn;
-                statum::__private::futures::future::join_all(
-                    self.__statum_items.iter().map(|__statum_item| {
-                        let __statum_fields = __statum_field_fn(__statum_item);
-                        __statum_item.into_machine()
-                            #field_builder_chain
-                            .#builder_method()
-                    })
-                ).await
             }
         }
     }
